+++ src/main/kotlin/ru/driics/aitrade/service/PromptBuilderService.kt
@@ -1,12 +1,9 @@
 package ru.driics.aitrade.service
 
-import com.fasterxml.jackson.module.kotlin.jacksonObjectMapper
 import org.slf4j.LoggerFactory
 import org.springframework.stereotype.Service
 import ru.driics.aitrade.config.PromptProperties
 import ru.driics.aitrade.model.MarketState
-import ru.driics.aitrade.model.Position
-import java.math.BigDecimal
 import java.time.LocalDateTime
 import java.time.ZoneId
 import java.time.format.DateTimeFormatter
@@ -14,10 +11,10 @@ import java.util.concurrent.TimeUnit
 
 @Service
 class PromptBuilderService(
-    private val promptProperties: PromptProperties
+    private val promptProperties: PromptProperties,
+    private val promptFormatterService: PromptFormatterService
 ) {
     private val log = LoggerFactory.getLogger(javaClass)
-    private val objectMapper = jacksonObjectMapper()
 
     fun buildPrompt(
         marketState: MarketState,
@@ -28,112 +25,11 @@ class PromptBuilderService(
         val currentTime = LocalDateTime.now(ZoneId.of("UTC"))
             .format(DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm:ss.SSSSSS"))
 
-        val builder = StringBuilder()
-
-        // Header
-        builder.append("It has been $minutesSinceStart minutes since you started trading. ")
-        builder.append("The current time is $currentTime and you've been invoked $invocationCount times. ")
-        builder.append("Below, we are providing you with a variety of state data, price data, and predictive signals so you can discover alpha. ")
-        builder.append("Below that is your current account information, value, performance, positions, etc. ")
-        builder.append("ALL OF THE PRICE OR SIGNAL DATA BELOW IS ORDERED: OLDEST → NEWEST\n")
-
-        // Timeframes note
-        builder.append("Timeframes note: Unless stated otherwise in a section title, intraday series are provided at 3‑minute intervals. ")
-        builder.append("If a coin uses a different interval, it is explicitly stated in that coin's section.\n\n")
-
-        // Current Market State
-        builder.append("CURRENT MARKET STATE FOR ALL COINS\n")
-
-        for ((symbol, data) in marketState.currencies) {
-            builder.append("ALL $symbol DATA\n")
-            builder.append("current_price = ${formatNumber(data.currentPrice)}, ")
-            builder.append("current_ema20 = ${formatNumber(data.currentEma20)}, ")
-            builder.append("current_macd = ${formatNumber(data.currentMacd)}, ")
-            builder.append("current_rsi (7 period) = ${formatNumber(data.currentRsi7)}\n")
-
-            // Open Interest and Funding Rate
-            if (data.openInterest != null || data.fundingRate != null) {
-                builder.append("In addition, here is the latest $symbol open interest and funding rate for perps (the instrument you are trading):\n")
-                data.openInterest?.let {
-                    builder.append("Open Interest: Latest: ${formatNumber(it)} Average: ${formatNumber(it)}\n")
-                }
-                data.fundingRate?.let {
-                    builder.append("Funding Rate: ${formatScientific(it)}\n")
-                }
-            }
-
-            // Intraday series
-            if (data.intradayPrices.isNotEmpty()) {
-                builder.append("Intraday series (by minute, oldest → latest):\n")
-                builder.append("Mid prices: ${formatNumberList(data.intradayPrices)}\n")
-
-                if (data.intradayEma20.isNotEmpty()) {
-                    builder.append("EMA indicators (20‑period): ${formatNumberList(data.intradayEma20)}\n")
-                }
-                if (data.intradayMacd.isNotEmpty()) {
-                    builder.append("MACD indicators: ${formatNumberList(data.intradayMacd)}\n")
-                }
-                if (data.intradayRsi7.isNotEmpty()) {
-                    builder.append("RSI indicators (7‑Period): ${formatNumberList(data.intradayRsi7)}\n")
-                }
-                if (data.intradayRsi14.isNotEmpty()) {
-                    builder.append("RSI indicators (14‑Period): ${formatNumberList(data.intradayRsi14)}\n")
-                }
-            }
-
-            // 4-hour context
-            if (data.ema20_4h != null || data.ema50_4h != null) {
-                builder.append("Longer‑term context (4‑hour timeframe):\n")
-
-                if (data.ema20_4h != null && data.ema50_4h != null) {
-                    builder.append("20‑Period EMA: ${formatNumber(data.ema20_4h)} vs. 50‑Period EMA: ${formatNumber(data.ema50_4h)}\n")
-                }
-
-                if (data.atr3_4h != null && data.atr14_4h != null) {
-                    builder.append("3‑Period ATR: ${formatNumber(data.atr3_4h)} vs. 14‑Period ATR: ${formatNumber(data.atr14_4h)}\n")
-                }
-
-                if (data.volume4h != null && data.avgVolume4h != null) {
-                    builder.append(
-                        "Current Volume: ${formatNumber(data.volume4h)} vs. Average Volume: ${
-                            formatNumber(
-                                data.avgVolume4h
-                            )
-                        }\n"
-                    )
-                }
-
-                if (data.macd4h.isNotEmpty()) {
-                    builder.append("MACD indicators: ${formatNumberList(data.macd4h)}\n")
-                }
-
-                if (data.rsi14_4h.isNotEmpty()) {
-                    builder.append("RSI indicators (14‑Period): ${formatNumberList(data.rsi14_4h)}\n")
-                }
-            }
-
-            builder.append("\n")
+        return buildString {
+            appendHeader(minutesSinceStart, currentTime, invocationCount)
+            appendMarketData(marketState)
+            appendAccountInfo(marketState)
         }
-
-        // Account Information
-        builder.append("HERE IS YOUR ACCOUNT INFORMATION & PERFORMANCE\n")
-        builder.append("Current Total Return (percent): ${formatPercent(marketState.account.totalReturn)}\n")
-        builder.append("Available Cash: ${formatNumber(marketState.account.availableCash)}\n")
-        builder.append("Current Account Value: ${formatNumber(marketState.account.accountValue)}\n")
-
-        // Positions
-        if (marketState.positions.isNotEmpty()) {
-            builder.append("Current live positions & performance: ")
-            builder.append(formatPositions(marketState.positions))
-            builder.append("\n")
-        }
-
-        // Sharpe Ratio
-        marketState.account.sharpeRatio?.let {
-            builder.append("Sharpe Ratio: ${formatNumber(it)}\n")
-        }
-
-        return builder.toString()
     }
 
     fun writePromptToFile(prompt: String): Boolean {
@@ -141,7 +37,6 @@ class PromptBuilderService(
             val file = java.io.File(promptProperties.outputPath)
             file.parentFile?.mkdirs()
 
-            // Atomic write: write to temp file first, then rename
             val tempFile = java.io.File("${promptProperties.outputPath}.tmp")
             tempFile.writeText(prompt)
             tempFile.renameTo(file)
@@ -158,76 +53,99 @@ class PromptBuilderService(
         println(prompt)
     }
 
-    private fun formatNumber(num: BigDecimal?): String {
-        if (num == null || num == BigDecimal.ZERO) return "0"
-
-        val abs = num.abs()
-        return when {
-            abs >= BigDecimal(1000000) -> num.setScale(2, java.math.RoundingMode.HALF_UP).toString()
-            abs >= BigDecimal(1) -> num.setScale(if (abs >= BigDecimal(100)) 0 else 2, java.math.RoundingMode.HALF_UP)
-                .toString()
-
-            abs >= BigDecimal(0.01) -> num.setScale(4, java.math.RoundingMode.HALF_UP).toString()
-            else -> num.setScale(8, java.math.RoundingMode.HALF_UP).toString()
-        }
+    private fun StringBuilder.appendHeader(minutes: Long, time: String, count: Long) {
+        append("It has been $minutes minutes since you started trading. ")
+        append("The current time is $time and you've been invoked $count times. ")
+        append("Below, we are providing you with a variety of state data, price data, and predictive signals so you can discover alpha. ")
+        append("Below that is your current account information, value, performance, positions, etc. ")
+        append("ALL OF THE PRICE OR SIGNAL DATA BELOW IS ORDERED: OLDEST → NEWEST\n")
+        append("Timeframes note: Unless stated otherwise in a section title, intraday series are provided at 3‑minute intervals. ")
+        append("If a coin uses a different interval, it is explicitly stated in that coin's section.\n\n")
     }
 
-    private fun formatScientific(num: BigDecimal?): String {
-        if (num == null) return "0"
-        return num.toPlainString()
-    }
+    private fun StringBuilder.appendMarketData(marketState: MarketState) {
+        append("CURRENT MARKET STATE FOR ALL COINS\n")
 
-    private fun formatPercent(num: BigDecimal?): String {
-        if (num == null || num == BigDecimal.ZERO) return "0%"
-        return "${num.setScale(2, java.math.RoundingMode.HALF_UP)}%"
-    }
+        for ((symbol, data) in marketState.currencies) {
+            append("ALL $symbol DATA\n")
+            append("current_price = ${promptFormatterService.formatNumber(data.currentPrice)}, ")
+            append("current_ema20 = ${promptFormatterService.formatNumber(data.currentEma20)}, ")
+            append("current_macd = ${promptFormatterService.formatNumber(data.currentMacd)}, ")
+            append("current_rsi (7 period) = ${promptFormatterService.formatNumber(data.currentRsi7)}\n")
 
-    private fun formatNumberList(numbers: List<BigDecimal>): String {
-        return "[${numbers.joinToString(", ") { formatNumber(it) }}]"
-    }
-
-    private fun formatPositions(positions: List<Position>): String {
-        if (positions.isEmpty()) return "{}"
-
-        return try {
-            // Convert to a simplified map structure for readability
-            val positionMaps = positions.map { pos ->
-                mutableMapOf<String, Any?>(
-                    "symbol" to pos.symbol,
-                    "quantity" to formatNumber(pos.quantity),
-                    "entry_price" to formatNumber(pos.entryPrice),
-                    "current_price" to formatNumber(pos.currentPrice),
-                    "unrealized_pnl" to formatNumber(pos.unrealizedPnl),
-                    "wait_for_fill" to pos.waitForFill
-                ).apply {
-                    pos.liquidationPrice?.let { put("liquidation_price", formatNumber(it)) }
-                    pos.leverage?.let { put("leverage", it) }
-                    pos.exitPlan?.let { plan ->
-                        put(
-                            "exit_plan", mapOf(
-                                "profit_target" to formatNumber(plan.profitTarget),
-                                "stop_loss" to formatNumber(plan.stopLoss),
-                                "invalidation_condition" to plan.invalidationCondition
-                            )
-                        )
-                    }
-                    pos.confidence?.let { put("confidence", formatNumber(it)) }
-                    pos.riskUsd?.let { put("risk_usd", formatNumber(it)) }
-                    pos.slOid?.let { put("sl_oid", it) }
-                    pos.tpOid?.let { put("tp_oid", it) }
-                    pos.entryOid?.let { put("entry_oid", it) }
-                    pos.notionalUsd?.let { put("notional_usd", formatNumber(it)) }
+            // Open Interest and Funding Rate
+            if (data.openInterest != null || data.fundingRate != null) {
+                append("In addition, here is the latest $symbol open interest and funding rate for perps (the instrument you are trading):\n")
+                data.openInterest?.let {
+                    append("Open Interest: Latest: ${promptFormatterService.formatNumber(it)} Average: ${promptFormatterService.formatNumber(it)}\n")
+                }
+                data.fundingRate?.let {
+                    append("Funding Rate: ${promptFormatterService.formatScientific(it)}\n")
                 }
             }
 
-            // Use single quotes for outer strings to match original format
-            objectMapper.writeValueAsString(positionMaps)
-                .replace("\"", "'")
-        } catch (e: Exception) {
-            log.error("Error formatting positions", e)
-            "{}"
+            // Intraday series
+            if (data.intradayPrices.isNotEmpty()) {
+                append("Intraday series (3‑minute intervals, oldest → latest):\n")
+                append("Mid prices: ${promptFormatterService.formatNumberList(data.intradayPrices)}\n")
+
+                if (data.intradayEma20.isNotEmpty()) {
+                    append("EMA indicators (20‑period): ${promptFormatterService.formatNumberList(data.intradayEma20)}\n")
+                }
+                if (data.intradayMacd.isNotEmpty()) {
+                    append("MACD indicators: ${promptFormatterService.formatNumberList(data.intradayMacd)}\n")
+                }
+                if (data.intradayRsi7.isNotEmpty()) {
+                    append("RSI indicators (7‑Period): ${promptFormatterService.formatNumberList(data.intradayRsi7)}\n")
+                }
+                if (data.intradayRsi14.isNotEmpty()) {
+                    append("RSI indicators (14‑Period): ${promptFormatterService.formatNumberList(data.intradayRsi14)}\n")
+                }
+            }
+
+            // 4-hour context
+            if (data.ema20_4h != null || data.ema50_4h != null) {
+                append("Longer‑term context (4‑hour timeframe):\n")
+
+                if (data.ema20_4h != null && data.ema50_4h != null) {
+                    append("20‑Period EMA: ${promptFormatterService.formatNumber(data.ema20_4h)} vs. 50‑Period EMA: ${promptFormatterService.formatNumber(data.ema50_4h)}\n")
+                }
+
+                if (data.atr3_4h != null && data.atr14_4h != null) {
+                    append("3‑Period ATR: ${promptFormatterService.formatNumber(data.atr3_4h)} vs. 14‑Period ATR: ${promptFormatterService.formatNumber(data.atr14_4h)}\n")
+                }
+
+                if (data.volume4h != null && data.avgVolume4h != null) {
+                    append("Current Volume: ${promptFormatterService.formatNumber(data.volume4h)} vs. Average Volume: ${promptFormatterService.formatNumber(data.avgVolume4h)}\n")
+                }
+
+                if (data.macd4h.isNotEmpty()) {
+                    append("MACD indicators: ${promptFormatterService.formatNumberList(data.macd4h)}\n")
+                }
+
+                if (data.rsi14_4h.isNotEmpty()) {
+                    append("RSI indicators (14‑Period): ${promptFormatterService.formatNumberList(data.rsi14_4h)}\n")
+                }
+            }
+
+            append("\n")
         }
     }
 
+    private fun StringBuilder.appendAccountInfo(marketState: MarketState) {
+        append("HERE IS YOUR ACCOUNT INFORMATION & PERFORMANCE\n")
+        append("Current Total Return (percent): ${promptFormatterService.formatPercent(marketState.account.totalReturn)}\n")
+        append("Available Cash: ${promptFormatterService.formatNumber(marketState.account.availableCash)}\n")
+        append("Current Account Value: ${promptFormatterService.formatNumber(marketState.account.accountValue)}\n")
 
+        if (marketState.positions.isNotEmpty()) {
+            append("Current live positions & performance: ")
+            append(promptFormatterService.formatPositions(marketState.positions))
+            append("\n")
+        }
+
+        marketState.account.sharpeRatio?.let {
+            append("Sharpe Ratio: ${promptFormatterService.formatNumber(it)}\n")
+        }
+    }
 }
